analytics-engine/requirements.txt
fastapi
uvicorn
pandas
openpyxl
requests
appwrite
python-dotenv
httpx
numpy
scipy
python-multipart
gender-guesser
